requirements.txt
django
djangorestframework
requests
Pillow
pytz
django-cors-headers
